+++ projects/swimlane/docspa-core/src/lib/modules/markdown-elements/md-print.component.ts
@@ -7,13 +7,12 @@ import { mergeMap, map } from 'rxjs/operators';
 import unified from 'unified';
 import markdown from 'remark-parse';
 import visit from 'unist-util-visit';
-import * as MDAST from 'mdast';
-import frontmatter from 'remark-frontmatter';
 import rehypeStringify from 'rehype-stringify';
 import remark2rehype from 'remark-rehype';
 import raw from 'rehype-raw';
 
-import { images } from '../../shared/links';
+import type * as mdast from 'mdast';
+
 import { LocationService } from '../../services/location.service';
 import { FetchService } from '../../services/fetch.service';
 import { SettingsService } from '../../services/settings.service';
@@ -21,8 +20,10 @@ import { RouterService } from '../../services/router.service';
 import { MarkdownService } from '../markdown/markdown.service';
 
 import { join, isAbsolutePath } from '../../shared/utils';
+import { images } from '../../shared/links';
 
-import type { VFile, Heading } from '../../vendor';
+import type { VFile } from '../../shared/vfile';
+import type { Heading } from '../../shared/ast';
 
 @Component({
   selector: 'docspa-print', // tslint:disable-line
@@ -102,8 +103,8 @@ export class MdPrintComponent implements OnInit {
 
     // TODO: move to service
     const fixLinks = () => {
-      return (tree: MDAST.Root, file: VFile) => {
-        return visit(tree, 'link', (node: MDAST.Link) => {
+      return (tree: mdast.Root, file: VFile) => {
+        return visit(tree, 'link', (node: mdast.Link) => {
           if (node && !isAbsolutePath(node.url)) {
             const url = locationService.prepareLink(node.url, file.history[0]).replace(/[\/#]/g, '--');
             node.url = `#${url}`;
@@ -115,7 +116,7 @@ export class MdPrintComponent implements OnInit {
 
     // TODO: move to service
     const fixIds = () => {
-      return (tree: MDAST.Root, file: VFile) => {
+      return (tree: mdast.Root, file: VFile) => {
         return visit(tree, 'heading', (node: Heading) => {
           if (node && node.data && node.data.hProperties && node.data.hProperties.id) {
             const id = locationService.prepareLink(`#${node.data.hProperties.id}`, file.history[0]).replace(/[\/#]/g, '--');
@@ -184,6 +185,7 @@ export class MdPrintComponent implements OnInit {
     });
   }
 
+  // TODO: use markdownService.getLinks
   private loadSummary(summary: string): Promise<string[]> {
     const vfile = this.locationService.pageToFile(summary);
     const fullPath = join(vfile.cwd, vfile.path);