+++ projects/swimlane/docspa-core/src/lib/services/router.service.ts
@@ -4,13 +4,13 @@ import { Router, ActivatedRouteSnapshot } from '@angular/router';
 import { NGXLogger } from 'ngx-logger';
 
 import { goExternal, isAbsolutePath } from '../shared/utils';
-import { getFullPath } from '../shared/vfile-utils';
+import { getFullPath } from '../shared/vfile';
 
 import { SettingsService } from './settings.service';
 import { FetchService } from './fetch.service';
 import { LocationService } from './location.service';
 
-import { VFile } from 'vfile';
+import { VFile } from '../shared/vfile';
 
 @Injectable({
   providedIn: 'root'