+++ projects/swimlane/docspa-core/src/lib/services/settings.service.ts
@@ -20,7 +20,7 @@ export class SettingsService {
   // nameLink = '';
   ext = '.md';
   notFoundPage = '_404.md';
-  maxPageCacheSize = 100;
+  maxPageCacheSize = 20;
 
   constructor(
     metaService: Meta,