+++ projects/swimlane/docspa-core/src/lib/shared/index.ts
@@ -1,5 +1,6 @@
 export * from './links';
 export * from './shortcodes';
 export * from './utils';
-export * from './vfile-utils';
+export * from './vfile';
 export * from './throttle';
+export * from './ast';